move einsum.py to einsum/__init__
